API Documentation

PlanetOrbits.orbit — Function

orbit(...)

Construct an orbit from the provided keyword arguments. Will automatically select a subclass of AbstractOrbit based on the information provided. This is a convenience function that is not type stable and should not be used in performance sensitive contexts. Instead, call one of the concrete constructors KepOrbit, VisualOrbit, or RadialVelocityOrbit directly. This function logs the kind of elements created so that it's easy to select the correct constructor.

Required arguments:

a: semi-major axis [AU]

M: mass of primary [M⊙]

Optional arguments:

τ: epoch of periastron passage at MJD=0 (default=0.0)

e: eccentricity (default=0)

ω: argument of periapsis rad

i: inclination [rad]

Ω: longitude of ascending node [rad]

plx: parallax [mas]; defines the distance to the primary

PlanetOrbits.VisualOrbit — Type

VisualOrbit(
    a, # semi-major axis [AU]
    e, # eccentricity
    i, # inclination [rad]
    ω, # argument of periapsis [rad]
    Ω, # longitude of ascending node [rad]
    τ, # epoch of periastron passage at MJD=0
    M, # mass of primary [M⊙]
    plx, # parallax [mas]; defines the distance to the primary
)

Represents the Keplerian elements of a secondary body orbiting a primary. Values can be specified by keyword argument or named tuple for convenience.

See also VisualOrbitDeg for a convenience constructor accepting units of degrees instead of radians for i, ω, and Ω.

PlanetOrbits.VisualOrbitDeg — Function

VisualOrbitDeg(a, e, i, ω, Ω, τ, M, plx)

A convenience function for constructing VisualOrbit where i, ω, and Ω are provided in units of degrees instead of radians.

PlanetOrbits.KepOrbitDeg — Function

KepOrbitDeg(a, e, i, ω, Ω, τ, M)

A convenience function for constructing KepOrbit where i, ω, and Ω are provided in units of degrees instead of radians.

PlanetOrbits.orbitsolve — Function

orbitsolve(elements, t, method=Auto())

Given a set of orbital elements with a time t in days, get the position and velocity of the secondary body (e.g. planet around a star).

This will output a struct that is a subtype of AbstractOrbitSolution which we can then query with raoff, decoff, radvel, etc.

You can also calculate those quanitities individually (see their docstrings) but if you need more than one, it is most efficient to save the orbit solution once.

Note: these calculations use the small angle approximation, so are only accurate when the star is much further way from the observer than the secondary is from the primary.

See also: orbitsolve_ν, orbitsolve_meananom, orbitsolve_eccanom, projectedseparation, raoff, decoff, radvel, propmotionanom.

PlanetOrbits.orbitsolve_ν — Function

orbitsolve_ν(elem, ν; EA)

Solve an orbit from a given true anomaly [rad]. See orbitsolve for the same function accepting a given time. Can optionally pass eccentric anomaly (EA) if already computed.

PlanetOrbits.orbitsolve_meananom — Function

orbitsolve_meananom(elements, MA)

Same as orbitsolve, but solves orbit for a given mean anomaly instead of time.

PlanetOrbits.orbitsolve_eccanom — Function

orbitsolve_eccanom(elements, EA)

Same as orbitsolve, but solves orbit for a given eccentric anomaly instead of time.

PlanetOrbits.OrbitSolutionVisual — Type

Represents a VisualOrbit evaluated to some position.

PlanetOrbits.OrbitSolutionKep — Type

Represents a KepOrbit evaluated to some position.

PlanetOrbits.OrbitSolutionRadialVelocity — Type

Represents a RadialVelocityOrbit evaluated to some position.

PlanetOrbits.posx — Function

posx(elem, t)

Get the offset [AU] from the primary body at the time t [days].

posx(elem, t)

Same as above, but from an instance of AbstractOrbitSolution.

PlanetOrbits.posy — Function

posy(elem, t)

Get the offset [AU] from the primary body at the time t [days].

posy(o)

Same as above, but from an instance of AbstractOrbitSolution.

PlanetOrbits.posz — Function

posz(elem, t)

Get the offset [AU] from the primary body at the time t [days].

posz(o)

Same as above, but from an instance of AbstractOrbitSolution.

PlanetOrbits.trueanom — Function

trueanom(orbit, t)

Get the true anomaly [radians] of the secondary at the time t [days].

trueanom(o)

Get the true anomaly [radians] of the secondary from an instance of AbstractOrbitSolution.

PlanetOrbits.eccanom — Function

eccanom(orbit, t)

Get the eccentric anomaly [radians] of the secondary at the time t [days].

eccanom(o)

Get the eccentric anomaly [radians] of the secondary from an instance of AbstractOrbitSolution.

PlanetOrbits.meananom — Function

meananom(orbit, t)

Get the mean anomaly [radians] of the secondary at the time t [days].

meananom(o)

Get the mean anomaly [radians] of the secondary from an instance of AbstractOrbitSolution.

PlanetOrbits.raoff — Function

raoff(elem, t)

Get the offset [mas] from the primary body in Right Ascension at the time t [days].

raoff(o)

Get the offset [mas] from the primary body in Right Ascension from an instance of AbstractOrbitSolution.

PlanetOrbits.decoff — Function

decoff(elem, t)

Get the offset [mas] from the primary body in Declination at the time t [days].

decoff(elem, t)

Get the offset [mas] from the primary body in Declination from an instance of AbstractOrbitSolution.

PlanetOrbits.radvel — Function

radvel(elem, t)

Get the radial velocity [m/s] of the secondary along the line of sight at the time t [days].

radvel(o)

Get the radial velocity [m/s] of the secondary along the line of sight from an instance of AbstractOrbitSolution.

radvel(elem, t, M_planet)

Get the radial velocity [m/s] of the primary along the line of sight at the time t [days]. The units of M_planet and elem.M must match.

radvel(o, M_planet)

Get the radial velocity [m/s] of the primary along the line of sight from an AbstractOrbitSolution. The units of M_planet and elem.M must match.

Missing docstring.

Missing docstring for accra. Check Documenter's build log for details.

Missing docstring.

Missing docstring for accdec. Check Documenter's build log for details.

PlanetOrbits.period — Function

period(elem)

Period of an orbit [days].

PlanetOrbits.distance — Function

distance(elem)

Distance to the system [pc].

PlanetOrbits.meanmotion — Function

meanmotion(elem)

Mean motion [rad/year].

PlanetOrbits.periastron — Function

periastron(elements, tref=58849)

Compute the MJD of periastron passage most recently after the reference epoch tref. N.B. mjd of 58849 = 2020-01-01

PlanetOrbits.posangle — Function

posangle(elem, t)

Calculate the position angle [rad] of the secondary about its primary from our perspective at the time t [days].

posangle(o)

Calculate the position angle [rad] of the secondary about its primary from our perspective from an instance of AbstractOrbitSolution.

posangle(elem, t, M_planet)

Calculate the position angle [rad] of the secondary about its primary from our perspective at the time t [days]. In this case only, the value of M_planet can be arbitrary.

posangle(o, M_planet)

Calculate the position angle [rad] of the primary from our perspective from an instance of AbstractOrbitSolution. In this case only, the value of M_planet can be arbitrary.

PlanetOrbits.projectedseparation — Function

projectedseparation(elem, t)

Calculate the projected separation [mas] of the secondary from its primary at the time t [days].

projectedseparation(o)

Calculate the projected separation [mas] of the secondary from its primary from an instance of AbstractOrbitSolution.

PlanetOrbits.Auto — Type

PlanetOrbits.Auto()

Automatic choice of Kepler solver algorithm. Currently defaults to PlanetOrbits.Markley()

PlanetOrbits.Markley — Type

PlanetOrbits.Markley()

Kepler solver implementation from AstroLib, based on Markley (1995) Celestial Mechanics and Dynamical Astronomy, 63, 101 (DOI:10.1007/BF00691917).

PlanetOrbits.Goat — Type

PlanetOrbits.Goat()

Kepler solver implementation from https://arxiv.org/abs/2103.15829 and https://github.com/oliverphilcox/Keplers-Goat-Herd

It is here for comparison purposes only. In general, Markley() is more performant and accurate.

PlanetOrbits.RootsMethod — Type

PlanetOrbits.RootsMethod(method::Roots.PlanetOrbits.Roots.AbstractUnivariateZeroMethod, kwargs...)

Wraps a root finding method from Roots.jl. Requires Roots to be loaded first. You can also pass keyword arguments that will be forwarded to Roots to control the tolerance.

Examples:

method = PlanetOrbits.RootsMethod(Roots.Newton())
method = PlanetOrbits.RootsMethod(Roots.Thukral5B())
method = PlanetOrbits.RootsMethod(Roots.Bisection())
method = PlanetOrbits.RootsMethod(Roots.A42())
method = PlanetOrbits.RootsMethod(Roots.Newton(), rtol=1e-3, atol=1e-3)