API Documentation

PlanetOrbits.AbstractOrbit — Type

AbstractOrbit

Represents a orbit. Contains all the information to calculate the location of a planet at a given time, true anomaly, eccentric anomaly, or mean anomaly. Different concrete implementations of AbstractOrbit contain varying amounts of information.

Basic information about the orbit can be queried using functions like period(orbit).

Orbits can be solved using functions like orbitsolve(orb).

See: RadialVelocityOrbit, KepOrbit, VisualOrbit

PlanetOrbits.AbstractOrbitSolution — Type

AbstractOrbitSolution

Represents the solution of an orbit. Contains all the information of an AbstractOrbit, plus information necessary to uniquely locate a planet.

The solution can be queried using a variety of functions such as radvel(solution).

The API for creating orbit solutions it not considered public as the fields may change between minor versions. Instead, create solutions only through the public orbitsolve and orbitsolve_... functions.

PlanetOrbits.Auto — Type

PlanetOrbits.Auto()

Automatic choice of Kepler solver algorithm. Currently defaults to PlanetOrbits.Markley()

PlanetOrbits.Goat — Type

PlanetOrbits.Goat()

Kepler solver implementation from https://arxiv.org/abs/2103.15829 and https://github.com/oliverphilcox/Keplers-Goat-Herd

It is here for comparison purposes only. In general, Markley() is more performant and accurate.

PlanetOrbits.KepOrbit — Type

KepOrbit(
    a, # semi-major axis [AU]
    e, # eccentricity
    i, # inclination [rad]
    ω, # argument of periapsis [rad]
    Ω, # longitude of ascending node [rad]
    τ, # epoch of periastron passage at MJD=0
    M, # mass of primary [M⊙]
)

Represents the Keplerian elements of a secondary body orbiting a primary. Values can be specified by keyword argument or named tuple for convenience.

See also KepOrbitDeg for a convenience constructor accepting units of degrees instead of radians for i, ω, and Ω.

PlanetOrbits.Markley — Type

PlanetOrbits.Markley()

Kepler solver implementation from AstroLib, based on Markley (1995) Celestial Mechanics and Dynamical Astronomy, 63, 101 (DOI:10.1007/BF00691917).

PlanetOrbits.OrbitSolutionKep — Type

Represents a KepOrbit evaluated to some position.

PlanetOrbits.OrbitSolutionRadialVelocity — Type

Represents a RadialVelocityOrbit evaluated to some position.

PlanetOrbits.OrbitSolutionVisual — Type

Represents a VisualOrbit evaluated to some position.

PlanetOrbits.RadialVelocityOrbit — Type

RadialVelocityOrbit(a, e, ω, τ, M)

Represents an orbit of a planet with only the information retrievable from radial velocity measurements. That is, without inclination, longitude of ascending node, or distance to the system.

PlanetOrbits.RootsMethod — Type

PlanetOrbits.RootsMethod(method::Roots.PlanetOrbits.Roots.AbstractUnivariateZeroMethod, kwargs...)

Wraps a root finding method from Roots.jl. Requires Roots to be loaded first. You can also pass keyword arguments that will be forwarded to Roots to control the tolerance.

Examples:

method = PlanetOrbits.RootsMethod(Roots.Newton())
method = PlanetOrbits.RootsMethod(Roots.Thukral5B())
method = PlanetOrbits.RootsMethod(Roots.Bisection())
method = PlanetOrbits.RootsMethod(Roots.A42())
method = PlanetOrbits.RootsMethod(Roots.Newton(), rtol=1e-3, atol=1e-3)

PlanetOrbits.VisualOrbit — Type

VisualOrbit(
    a, # semi-major axis [AU]
    e, # eccentricity
    i, # inclination [rad]
    ω, # argument of periapsis [rad]
    Ω, # longitude of ascending node [rad]
    τ, # epoch of periastron passage at MJD=0
    M, # mass of primary [M⊙]
    plx, # parallax [mas]; defines the distance to the primary
)

Represents the Keplerian elements of a secondary body orbiting a primary. Values can be specified by keyword argument or named tuple for convenience.

See also VisualOrbitDeg for a convenience constructor accepting units of degrees instead of radians for i, ω, and Ω.

PlanetOrbits.KepOrbitDeg — Method

KepOrbitDeg(a, e, i, ω, Ω, τ, M)

A convenience function for constructing KepOrbit where i, ω, and Ω are provided in units of degrees instead of radians.

PlanetOrbits.VisualOrbitDeg — Method

VisualOrbitDeg(a, e, i, ω, Ω, τ, M, plx)

A convenience function for constructing VisualOrbit where i, ω, and Ω are provided in units of degrees instead of radians.

PlanetOrbits.acceleration — Method

acceleration(elem, t)

Get the instantaneous acceleration [mas/year^2] of the secondary at the time t [days].

acceleration(o)

Get the instantaneous acceleration [mas/year^2] of the secondary from an instance of AbstractOrbitSolution.

acceleration(elem, t, M_planet)

Get the instantaneous acceleration [mas/year^2] of the primary in at the time t [days]. The units of M_planet and elem.M must match.

acceleration(o)

Get the instantaneous acceleration [mas/year^2] of the primary from an instance of AbstractOrbitSolution. The units of M_planet and elem.M must match.

PlanetOrbits.astuple — Method

astuple(elements)

Return the parameters of a KepOrbit value as a tuple.

PlanetOrbits.astuple — Method

astuple(elements)

Return the parameters of a VisualOrbit value as a tuple.

PlanetOrbits.decoff — Function

decoff(elem, t)

Get the offset [mas] from the primary body in Declination at the time t [days].

decoff(elem, t)

Get the offset [mas] from the primary body in Declination from an instance of AbstractOrbitSolution.

PlanetOrbits.distance — Function

distance(elem)

Distance to the system [pc].

PlanetOrbits.eccanom — Method

eccanom(orbit, t)

Get the eccentric anomaly [radians] of the secondary at the time t [days].

eccanom(o)

Get the eccentric anomaly [radians] of the secondary from an instance of AbstractOrbitSolution.

PlanetOrbits.meananom — Method

meananom(orbit, t)

Get the mean anomaly [radians] of the secondary at the time t [days].

meananom(o)

Get the mean anomaly [radians] of the secondary from an instance of AbstractOrbitSolution.

PlanetOrbits.meanmotion — Function

meanmotion(elem)

Mean motion [rad/year].

PlanetOrbits.mjd — Method

mjd("2020-01-01")

Get the modfied julian day of a date, or in general a UTC timestamp.

PlanetOrbits.mjd — Method

mjd()

Get the current modified julian day of right now.

PlanetOrbits.mjd2date — Method

mjd2date(modified_julian)

Get a Date value from a modfied julian day, rounded to closest day

Examples

julia> mjd2date(59160.8)
2020-11-08

PlanetOrbits.orbit — Method

orbit(...)

Construct an orbit from the provided keyword arguments. Will automatically select a subclass of AbstractOrbit based on the information provided. This is a convenience function that is not type stable and should not be used in performance sensitive contexts. Instead, call one of the concrete constructors KepOrbit, VisualOrbit, or RadialVelocityOrbit directly. This function logs the kind of elements created so that it's easy to select the correct constructor.

Required arguments:

a: semi-major axis [AU]

M: mass of primary [M⊙]

Optional arguments:

τ: epoch of periastron passage at MJD=0, default=0

e: eccentricity, default=0

ω: argument of periapsis [rad], default=0

i: inclination [rad]

Ω: longitude of ascending node [rad]

plx: parallax [mas]; defines the distance to the primary

PlanetOrbits.orbitsolve — Function

orbitsolve(elements, t, method=Auto())

Given a set of orbital elements with a time t in days, get the position and velocity of the secondary body (e.g. planet around a star).

This will output a struct that is a subtype of AbstractOrbitSolution which we can then query with raoff, decoff, radvel, etc.

You can also calculate those quanitities individually (see their docstrings) but if you need more than one, it is most efficient to save the orbit solution once.

Note: these calculations use the small angle approximation, so are only accurate when the star is much further way from the observer than the secondary is from the primary.

See also: orbitsolve_ν, orbitsolve_meananom, orbitsolve_eccanom, projectedseparation, raoff, decoff, radvel, propmotionanom.

PlanetOrbits.orbitsolve_eccanom — Method

orbitsolve_eccanom(elements, EA)

Same as orbitsolve, but solves orbit for a given eccentric anomaly instead of time.

PlanetOrbits.orbitsolve_meananom — Method

orbitsolve_meananom(elements, MA)

Same as orbitsolve, but solves orbit for a given mean anomaly instead of time.

PlanetOrbits.orbitsolve_ν — Function

orbitsolve_ν(elem, ν)

Solve a keplerian orbit from a given true anomaly [rad]. See orbitsolve for the same function accepting a given time.

PlanetOrbits.orbitsolve_ν — Function

orbitsolve_ν(elem, ν, EA)

Solve an orbit from a given true anomaly [rad]. See orbitsolve for the same function accepting a given time. Can optionally pass eccentric anomaly (EA) if already computed.

PlanetOrbits.periastron — Function

periastron(elements, tref=58849)

Compute the MJD of periastron passage most recently after the reference epoch tref. N.B. mjd of 58849 = 2020-01-01

PlanetOrbits.period — Function

period(elem)

Period of an orbit [days].

PlanetOrbits.posangle — Method

posangle(elem, t)

Calculate the position angle [rad] of the secondary about its primary from our perspective at the time t [days].

posangle(o)

Calculate the position angle [rad] of the secondary about its primary from our perspective from an instance of AbstractOrbitSolution.

posangle(elem, t, M_planet)

Calculate the position angle [rad] of the secondary about its primary from our perspective at the time t [days]. In this case only, the value of M_planet can be arbitrary.

posangle(o, M_planet)

Calculate the position angle [rad] of the primary from our perspective from an instance of AbstractOrbitSolution. In this case only, the value of M_planet can be arbitrary.

PlanetOrbits.posx — Function

posx(elem, t)

Get the offset [AU] from the primary body at the time t [days].

posx(elem, t)

Same as above, but from an instance of AbstractOrbitSolution.

PlanetOrbits.posy — Function

posy(elem, t)

Get the offset [AU] from the primary body at the time t [days].

posy(o)

Same as above, but from an instance of AbstractOrbitSolution.

PlanetOrbits.posz — Function

posz(elem, t)

Get the offset [AU] from the primary body at the time t [days].

posz(o)

Same as above, but from an instance of AbstractOrbitSolution.

PlanetOrbits.projectedseparation — Method

projectedseparation(elem, t)

Calculate the projected separation [mas] of the secondary from its primary at the time t [days].

projectedseparation(o)

Calculate the projected separation [mas] of the secondary from its primary from an instance of AbstractOrbitSolution.

PlanetOrbits.propmotionanom — Method

propmotionanom(elem, t)

Get the instantaneous proper motion anomaly [mas/year] of the secondary at the time t [days].

propmotionanom(o)

Get the instantaneous proper motion anomaly [mas/year] of the secondary from an instance of AbstractOrbitSolution.

propmotionanom(elem, t, M_planet)

Get the instantaneous proper motion anomaly [mas/year] of the primary in at the time t [days]. The units of M_planet and elem.M must match.

propmotionanom(o, M_planet)

Same as above, but from an orbit solution.

PlanetOrbits.radvel — Function

radvel(elem, t)

Get the radial velocity [m/s] of the secondary along the line of sight at the time t [days].

radvel(o)

Get the radial velocity [m/s] of the secondary along the line of sight from an instance of AbstractOrbitSolution.

radvel(elem, t, M_planet)

Get the radial velocity [m/s] of the primary along the line of sight at the time t [days]. The units of M_planet and elem.M must match.

radvel(o, M_planet)

Get the radial velocity [m/s] of the primary along the line of sight from an AbstractOrbitSolution. The units of M_planet and elem.M must match.

PlanetOrbits.raoff — Function

raoff(elem, t)

Get the offset [mas] from the primary body in Right Ascension at the time t [days].

raoff(o)

Get the offset [mas] from the primary body in Right Ascension from an instance of AbstractOrbitSolution.

PlanetOrbits.semiamplitude — Function

semiamplitude(elem)

Radial velocity semiamplitude [m/s].

PlanetOrbits.trueanom — Method

trueanom(orbit, t)

Get the true anomaly [radians] of the secondary at the time t [days].

trueanom(o)

Get the true anomaly [radians] of the secondary from an instance of AbstractOrbitSolution.

PlanetOrbits.years2mjd — Method

years2mjd()

Convert from fractional years (e.g. 1995.25) into modified julian date.